+++ infra/app.py
@@ -19,6 +19,7 @@ import aws_cdk as cdk  # noqa: E402
 
 from infra.stacks.data import DataStack  # noqa: E402
 from infra.stacks.identity import IdentityStack  # noqa: E402
+from infra.stacks.lti import LtiStack  # noqa: E402
 
 app = cdk.App()
 
@@ -30,5 +31,6 @@ env = cdk.Environment(
 
 IdentityStack(app, "agg-identity", env=env)
 DataStack(app, "agg-data", env=env)
+LtiStack(app, "agg-lti", env=env)
 
 app.synth()